Move logos to design system
