+++ homepage/homepage/app/docs/guide.mdx
@@ -1,7 +1,6 @@
 import { HeroHeader } from "gcmp-design-system/src/app/components/molecules/HeroHeader";
-import { JazzLogo } from "@/components/logos";
 import { CodeGroup } from "@/components/forMdx";
-import { ComingSoonBadge } from 'gcmp-design-system/src/app/components/atoms/ComingSoonBadge'
+import { JazzLogo } from 'gcmp-design-system/src/app/components/atoms/logos/JazzLogo'
 
 <div className="not-prose">
 <HeroHeader
